Exercise 5.11 - Add confirmation dialog for delete button

+++ part5/bloglist-frontend/src/App.jsx
@@ -79,9 +79,11 @@ const App = () => {
 
   const removeBlog = (blogToRemove) => {
     try {
-      blogService.remove(blogToRemove)
-      updateBlogList(blogs.filter(blog => blog.id !== blogToRemove.id))
-      showNotification(`Removed ${blogToRemove.title}`, false)
+      if(window.confirm(`Remove blog ${blogToRemove.title} by ${blogToRemove.author}?`)) {
+        blogService.remove(blogToRemove)
+        updateBlogList(blogs.filter(blog => blog.id !== blogToRemove.id))
+        showNotification(`Removed ${blogToRemove.title}`, false)
+      }
     } catch(exception) {
       showNotification(`Error removing blog: ${exception.message}`, true)
     }
